+++ Day05.cs
@@ -9,7 +9,7 @@ namespace aoc2019
     {
         public static void Run()
         {
-            var input = File.ReadAllLines(Path.Combine("input", "day05.txt")).SelectMany(l => l.Split(',')).Select(Int32.Parse).ToList();
+            var input = File.ReadAllLines(Path.Combine("input", "day05.txt")).SelectMany(l => l.Split(',')).Select(Int64.Parse).ToList();
 
             var res1 = CheckTestsAndGetDiagnosticCode(input, 1);
             Console.WriteLine($"Day05 - part1 - result: {res1}");
@@ -18,7 +18,7 @@ namespace aoc2019
             Console.WriteLine($"Day05 - part2 - result: {res2}");
         }
 
-        public static int CheckTestsAndGetDiagnosticCode(List<int> program, int input)
+        public static long CheckTestsAndGetDiagnosticCode(List<long> program, long input)
         {
             var computer = new IntComputer();
             computer.Run(program, new[] {input});
